Give daemon errors their own file
